Fix check marker not appearing after undoing moves
A bug which was caused by not running the createTilesUnderThreat
Function again after updating the board with the undone move - therefore
the game was looking for a checked tile on the old board, finding no
corresponding tile and therefore not correctly showing that somebody was
in check. Now it should be fixed.