Check player color upon confirmation
If player color is the opposite of the color who originally requested
the undo of the move the button actually activates the method to undo
the last move. If not, the button does nothing. Doing nothing is not
perfect - it should change the looks/not be available for clicking at
all. Working on it in the next commit.