Fix the bug
This should fix the bug. The bug was caused by calling the create
threats function when undoing with the rook of the same color basically
running the threat function for and by the player which has just done the castle. This creates the threats of the own pieces, which gets detected as check by the algorithm, which determines if a castle can be done, preventing castling again.