Remove bug... again?
The bug was still there - only visible for the white player though... it
the solution I deployed first was not cutting it, because there was
another call to the threat function from the undo handler itself. And
the undo handler did not clean any marks left by the updateBoard
function, that's why I got the bug in the first place. Now it cleans the
board markers before applying the new check markers and now it should
definetely work once and for all.