Add functionality: turn baseline pawn into figure
This is the only commit in this branch. The branch adds a core
functionality to the game of chess. It allows players to change their
pawn into a figure of their choice once it reaches the enemies baseline.
The game will pause and not accept any more inputs until the choice has
been done. The move is affected by the undo just like all the other
moves. Upon undoing the newly created piece vanishes and makes place for
whatever came before it - be it an empty tile or a captured enemy piece.