Add secure-server-methods
This is the only commit in this branch. It features a major rewrite of
all the server methods, adding several new ones and changing or deleting
the unsafe ones. Now the game should be a lot more safe. It also
separates the client from the server side logic a lot - which is a major
good part of this refactoring.