Add Connect-Disconnect Display
This adds two small dots nexto the players names in every game which get
colored green or red depending on the connection status of the players.
It checks the connection by sending an update signal every second to the
server which actualizes a timestamp which the symbols then compare with
the actual time - if the difference is greater than the rate with which
the active component updates the timestamp the player gets flagged as
offline.